loki-fundamentals/greenhouse/requirements.txt
bidict==0.23.1
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-engineio==4.9.0
python-socketio==5.11.2
simple-websocket==1.0.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
wsproto==1.2.0
psycopg2
requests
fluent-logger
